requirements.txt
torch>=2.0.0
triton>=2.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pytest>=7.0.0
pyyaml>=6.0.0
matplotlib>=3.7.0
numpy>=1.24.0
